--- examples/python/auto-tracing/app/db/__init__.py
@@ -1,3 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
