+++ examples/python/auto-tracing/app/server/__init__.py
@@ -1,7 +1,20 @@
-from flask import Flask, jsonify, render_template, request
+from datetime import datetime
 
-from app.db import db
-from app.db.models import Calculation
+from flask import Flask, jsonify, render_template, request
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+
+class Calculation(db.Model):
+    """Store calculation information."""
+
+    id = db.Column(db.Integer, primary_key=True)
+    type = db.Column(db.String(50), nullable=False)
+    input_data = db.Column(db.String(200), nullable=False)
+    result = db.Column(db.Float, nullable=False)
+    timestamp = db.Column(db.DateTime, default=datetime.utcnow)
+
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///financial_calculator.db'
